pyproject.toml
[project]
name = "rag-ai-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cohere>=5.20.7",
    "datasets>=4.8.3",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.135.1",
    "langchain>=1.2.13",
    "langchain-cohere>=0.5.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.11",
    "pydantic>=2.12.5",
    "pymupdf>=1.27.2.2",
    "python-dotenv>=1.2.2",
    "ragas>=0.4.3",
    "rank-bm25>=0.2.2",
    "uvicorn>=0.42.0",
]
